Refactor code structure; rename files, update imports, and improve widget organization
