+++ lib/core/helpers/bloc_observer.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_print
+
 import 'package:bloc/bloc.dart';
 
 class MyBlocObserver extends BlocObserver {
