+++ lib/features/search_page/controller/cubit/search_product_cubit.dart
@@ -1,6 +1,4 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:sadeem_tech_intern/core/networking/api_error_handler.dart';
 import 'package:sadeem_tech_intern/features/search_page/controller/cubit/search_product_state.dart';
 import 'package:sadeem_tech_intern/features/search_page/data/repos/search_product_repo.dart';
