+++ lib/main.dart
@@ -4,10 +4,10 @@ import 'package:sadeem_tech_intern/core/constants/services.dart';
 import 'package:sadeem_tech_intern/core/di/dependancy.dart';
 import 'package:sadeem_tech_intern/core/routing/app_router.dart';
 import 'package:sadeem_tech_intern/features/cart_page/controller/cubit/cart_cubit.dart';
-import 'package:sadeem_tech_intern/myApp.dart';
+import 'package:sadeem_tech_intern/my_app.dart';
 
 void main() async {
-  servicesLocator();
+ await servicesLocator();
   runApp(
     BlocProvider(
       create: (context) => getIt<CartCubit>(),
